keep sigtem and set early_stop before isalive check

+++ ts/nni_manager/training_service/local/localTrainingService.ts
@@ -236,12 +236,21 @@ class LocalTrainingService implements TrainingService {
 
             return Promise.resolve();
         }
-        tkill(trialJob.pid, 'SIGKILL', (err) => {
-            if (err) {
-                this.log.error(`kill trial job error: ${err}`);
-            }
-        });
+        tkill(trialJob.pid, 'SIGTERM');
         this.setTrialJobStatus(trialJob, getJobCancelStatus(isEarlyStopped));
+
+        const startTime = Date.now();
+        while(await isAlive(trialJob.pid)) {
+            if (Date.now() - startTime > 4999) {
+                tkill(trialJob.pid, 'SIGKILL', (err) => {
+                    if (err) {
+                        this.log.error(`kill trial job error: ${err}`);
+                    }
+                });
+                break;
+            }
+            await delay(500);
+        }
 
         return Promise.resolve();
     }
